# admin/docs/openapi.yaml
openapi: 3.1.0
info:
  title: 팝콘톡 Admin API
  version: 0.1.0
  license:
    name: Proprietary
  description: |
    팝콘톡 관리자 콘솔이 쓰는 API입니다.
    화면 10개가 실제로 호출하는 것만 담았고, 화면에 없는 기능은 넣지 않았습니다.

    ## 작성 근거

    - **영화 도메인**은 실제 DB(`dev` 스키마)의 컬럼명·타입을 그대로 씁니다.
      화면에서 이름을 바꾸면 매핑 층이 하나 더 생기므로 그러지 않았습니다.
    - **회원 도메인**은 아직 테이블이 없습니다. `docs/user-domain-draft.sql`의
      초안을 전제로 작성했으므로, 스키마가 확정되면 함께 조정해야 합니다.
    - 조회 응답은 `dev.popcorn_movies_service` 뷰를 그대로 내려주는 형태입니다.
      이 뷰가 `is_embedded`와 `media`를 미리 붙여줍니다.

    ## 규약

    | 항목 | 규칙 |
    |---|---|
    | 목록 | `page`(1부터) · `size`(기본 10) · 응답에 `page`·`size`·`total`·`total_pages` |
    | 날짜 | 날짜는 `YYYY-MM-DD`, 시각은 `YYYY-MM-DD HH:mm:ss` (KST) |
    | 기간 필터 | `*_from` · `*_to`, 양끝 포함. 한쪽만 주면 그쪽만 제한 |
    | 정렬 | `sort=필드:asc|desc`. 기본값은 엔드포인트마다 명시 |
    | 부분 수정 | `PATCH` + 바꿀 필드만. 빈 객체는 400 |
    | 오류 | RFC 9457 `application/problem+json` |

    ## 인증

    > **이번 범위에서 관리자 로그인은 만들지 않기로 했습니다.**
    > `/auth/*` 세 개는 목표 설계로 남겨둔 것이며 이번 구현 대상이 아닙니다.
    > 그동안 API를 무엇으로 보호할지, 감사 기록의 "누가"를 무엇으로 채울지는
    > 아래 *로그인 없이 가는 동안* 을 보세요.

    `/auth/login`·`/auth/refresh`를 제외한 모든 요청에
    `Authorization: Bearer <JWT>`가 필요합니다.
    토큰에는 관리자 `id`와 `role`이 들어갑니다. 감사 로그의 "누가"는 서버가
    토큰에서 읽어 채웁니다 — 클라이언트가 보내는 값을 믿지 않습니다.

    | | |
    |---|---|
    | 액세스 토큰 | 15분. 응답 본문으로 내려주고 **메모리에만** 보관 |
    | 리프레시 토큰 | 14일. `httpOnly` · `Secure` · `SameSite=Strict` 쿠키, 갱신 시 회전 |

    액세스 토큰을 `localStorage`에 넣으면 XSS 한 번에 털립니다.

...

    ## 회원과 관리자는 한 테이블입니다

...

...

...

...

...

...

...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /movies:
    get:
      operationId: listMovies
      tags: [movies]
      summary: 영화 목록
      description: |
        서비스 영화 화면과 인증완료 영화 화면이 같은 엔드포인트를 씁니다.
        인증완료 화면은 `approval_status=APPROVED`를 붙여 호출합니다.
      parameters:
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/Size' }
        - name: approval_status
          in: query
          description: 여러 개면 콤마로 구분
          schema: { type: string, example: 'PENDING,APPROVED' }
        - name: service_status
          in: query
          schema: { $ref: '#/components/schemas/MovieServiceStatus' }
        - name: genre
          in: query
          description: 장르 정확 일치. `genres` 배열에 포함되면 매칭
          schema: { type: string, example: 드라마 }
        - name: q
          in: query
          description: 제목·감독·배우 부분 일치
          schema: { type: string }
        - name: synced_from
          in: query
          description: 수집일시 시작
          schema: { type: string, format: date }
        - name: synced_to
          in: query
          schema: { type: string, format: date }
        - name: approved_from
          in: query
          description: 인증일시 시작. 인증완료 화면의 기간 필터
          schema: { type: string, format: date }
        - name: approved_to
          in: query
          schema: { type: string, format: date }
        - name: sort
          in: query
          schema:
            type: string
            default: release_date:desc
            enum: [release_date:desc, release_date:asc, created_at:desc, pop_talk_score:desc]
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PageMeta'
                  - type: object
                    required: [items, summary]
                    properties:
                      items:
                        type: array
                        items: { $ref: '#/components/schemas/Movie' }
                      summary:
                        $ref: '#/components/schemas/MovieSummary'
              example:
                page: 1
                size: 10
                total: 5985
                total_pages: 599
                items:
                  - id: 1
                    kofic_movie_cd: '20254904'
                    kmdb_id: K41667
                    kmdb_matched: true
                    title_ko: 가능한 사랑
                    title_en: Possible Love
                    release_date: '2025-09-23'
                    production_year: 2025
                    runtime_minutes: 164
                    production_status: 개봉예정
                    production_countries: [한국]
                    representative_country: 한국
                    genres: [드라마]
                    directors: [이창동]
                    actors: [전도연, 설경구, 조인성, 조여정]
                    viewing_grade: 청소년관람불가
                    poster_url: 'http://file.koreafilm.or.kr/thm/02/99/19/51/tn_DPK026223.jpg'
                    service_status: PUBLISHED
                    approval_status: APPROVED
                    approved_by: initial-dataset
                    approved_at: '2026-08-07 10:46:13'
                    source_system: KOFIC_KMDB
                    source_hash: 9f2c1a7e4b83d05c6e1f8a37b2d94c05e7a13f6b8c2d94e05a7f1c3b6d82e940
                    source_synced_at: '2026-08-07 10:46:13'
                    created_at: '2026-08-07 10:46:13'
                    updated_at: '2026-08-07 10:46:13'
                    is_embedded: false
                    media:
                      - { type: POSTER, url: 'http://file.koreafilm.or.kr/thm/02/99/19/51/tn_DPK026223.jpg', order: 0, primary: true }
                      - { type: STILL, url: 'http://file.koreafilm.or.kr/thm/01/copy/00/78/48/tn_DST953024.jpg', order: 0, primary: false }
                    pop_talk_score: null
                    categories: []
                summary: { total: 5985, published: 5985, pending: 0, approved: 5985, rejected: 0 }

        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
  /reviews:
    get:
      operationId: listReviews
      tags: [reviews]
      summary: 감상평 목록
      parameters:
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/Size' }
        - name: status
          in: query
          schema: { $ref: '#/components/schemas/ReviewStatus' }
        - name: member_id
          in: query
          description: 회원 상세의 감상평 탭이 사용
          schema: { type: integer, format: int64 }
        - name: q
          in: query
          description: 작성자·영화·내용 부분 일치
          schema: { type: string }
        - name: created_from
          in: query
          schema: { type: string, format: date }
        - name: created_to
          in: query
          schema: { type: string, format: date }
        - name: sort
          in: query
          schema: { type: string, default: created_at:desc }
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PageMeta'
                  - type: object
                    properties:
                      items:
                        type: array
                        items: { $ref: '#/components/schemas/Review' }
                      summary:
                        $ref: '#/components/schemas/ReviewSummary'

        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /reviews/{reviewId}/status:
    parameters:
      - name: reviewId
        in: path
        required: true
        schema: { type: integer, format: int64 }
    put:
      operationId: setReviewsStatus
      tags: [reviews]
      summary: 감상평 상태 변경 (정상 / 숨김 / 삭제)
      description: |
        `DELETED`는 물리 삭제가 아니라 상태 변경입니다.
        처리 이력을 남겨야 하므로 행을 지우지 않습니다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status: { $ref: '#/components/schemas/ReviewStatus' }
                reason: { type: string, description: 숨김·삭제 사유 }
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Review' }

        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  # ────────────────────────────────────────────────────────────────
  # 회원
  # ────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Page:
      name: page
      in: query
      description: 1부터 시작
      schema: { type: integer, minimum: 1, default: 1 }
    Size:
      name: size
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 10 }
...
  responses:
...
    Unauthorized:
      description: 인증 필요
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Forbidden:
      description: 권한 없음
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
  schemas:
...
    PageMeta:
      type: object
      required: [page, size, total, total_pages]
      properties:
        page: { type: integer, example: 1 }
        size: { type: integer, example: 10 }
        total: { type: integer, example: 5985 }
        total_pages: { type: integer, example: 599 }
    Problem:
      type: object
      description: RFC 9457
      required: [type, title, status]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string, format: uri }
        errors:
          type: array
          description: 필드 단위 검증 실패
          items:
            type: object
            properties:
              field: { type: string }
              message: { type: string }

    # ── enum (dev 스키마와 1:1) ─────────────────────────────
    MovieApprovalStatus:
      type: string
      enum: [PENDING, APPROVED, REJECTED]
      description: 인증대기 · 인증완료 · 반려
    MovieServiceStatus:
      type: string
      enum: [DRAFT, PUBLISHED, HIDDEN]
      description: 준비중 · 노출중 · 숨김
...
    MovieMediaType:
      type: string
      enum: [POSTER, STILL]
...
    ReviewStatus:
      type: string
      enum: [NORMAL, HIDDEN, DELETED]
...
    Movie:
      type: object
      description: '`dev.popcorn_movies_service` 뷰의 한 행'
      required:
        [id, kofic_movie_cd, kmdb_matched, title_ko, release_date,
         genres, directors, actors, service_status, approval_status]
      properties:
        id: { type: integer, format: int64 }
        kofic_movie_cd:
          type: string
          description: KOFIC 영화 코드. 필수·유니크한 원천 키
        kmdb_id: { type: [string, 'null'] }
        kmdb_matched:
          type: boolean
          description: false면 포스터·줄거리가 비어 있을 가능성이 높음 (실 데이터의 6%)
        title_ko: { type: string }
        title_en: { type: [string, 'null'] }
        title_original: { type: [string, 'null'] }
        release_date: { type: string, format: date }
        production_year: { type: [integer, 'null'] }
        runtime_minutes: { type: [integer, 'null'] }
        movie_type: { type: [string, 'null'] }
        production_status:
          type: [string, 'null']
          description: '개봉 · 개봉예정'
        production_countries: { type: array, items: { type: string } }
        representative_country: { type: [string, 'null'] }
        genres: { type: array, items: { type: string } }
        representative_genre: { type: [string, 'null'] }
        directors: { type: array, items: { type: string } }
        director_names_en: { type: array, items: { type: string } }
        actors: { type: array, items: { type: string } }
        actor_roles: { type: array, items: { type: string } }
        production_companies: { type: array, items: { type: string } }
        viewing_grade:
          type: [string, 'null']
          description: KOFIC 원문 그대로. 목록에서는 축약해 보여줌
        poster_url:
          type: [string, 'null']
          description: 비어 있으면 `media`의 대표 포스터를 씀 (그런 영화가 35편)
        plot: { type: [string, 'null'] }
        source_keywords: { type: array, items: { type: string } }
        service_status: { $ref: '#/components/schemas/MovieServiceStatus' }
        approval_status: { $ref: '#/components/schemas/MovieApprovalStatus' }
        approved_by: { type: [string, 'null'] }
        approved_at: { type: [string, 'null'], example: '2026-08-05 14:20:33' }
        rejection_reason:
          type: [string, 'null']
          description: '`approval_status=REJECTED`면 비어 있을 수 없음'
        source_system: { type: string, example: KOFIC_KMDB }
        source_hash:
          type: string
          description: 원본 행 변경 감지를 위한 SHA-256. 배치가 채우며 admin은 읽기만 합니다
        source_synced_at: { type: string, example: '2026-08-07 10:46:13' }
        created_at: { type: string }
        updated_at: { type: string }
        is_embedded:
          type: boolean
          description: |
            뷰가 계산합니다. `PROFILE` 임베딩이 `READY`이고 벡터가 실제로 있는지.

            **생성은 CLOVA Studio Embedding v2가 하는데 `embedding_model` 컬럼에는
            `bge-m3`가 기록됩니다.** 배치 코드(`embedding/worker.py`)가 그렇게 넣고,
            뷰도 `embedding_model = 'bge-m3'`로 거릅니다. 이름과 실제가 다르니
            모델을 바꿀 때 이 값도 함께 봐야 합니다.
        media:
          type: array
          description: 포스터·스틸 전부. 영화당 평균 8.3개, 최대 53개
          items: { $ref: '#/components/schemas/MovieMediaEntry' }
        pop_talk_score:
          type: [number, 'null']
          description: '**DB에 아직 없는 컬럼.** 추가 예정'
        categories:
          type: array
          description: '**DB에 아직 테이블이 없음.** 추가 예정'
          items: { type: string }
    MovieMediaEntry:
      type: object
      required: [type, url, order, primary]
      properties:
        type: { $ref: '#/components/schemas/MovieMediaType' }
        url: { type: string, format: uri }
        order: { type: integer }
        primary: { type: boolean }
...
    MovieSummary:
      type: object
      required: [total, published, pending, approved, rejected]
      description: 서비스 영화 화면의 요약 카드 4장
      properties:
        total: { type: integer }
        published: { type: integer, description: 노출 중 }
        pending: { type: integer }
        approved: { type: integer }
        rejected: { type: integer }

    # ── 카테고리 ──────────────────────────────────────────
...
    Review:
      type: object
      required: [id, member_id, author, movie_id, movie, rating, content, status, created_at]
      properties:
        id: { type: string, format: uuid }
        member_id: { type: string, format: uuid }
        author: { type: string, description: 회원 닉네임 }
        source_system:
          type: [string, 'null']
          example: naver_movie
          description: |
            감상평의 출처. `dev.reviews.source_system`입니다.

            **회원이 앱에서 직접 쓴 감상평은 `null`입니다.** 마이그레이션 008이
            `source_system`·`source_user_key`·`source_review_key` 셋을 CHECK로
            묶어 두어, 전부 차 있거나 전부 비어 있거나 둘 중 하나입니다.

            수집분과 회원 감상평이 실제로 섞여 있으므로 `null`을 특정 출처로
            메우지 마세요. admin은 이 값으로 목록의 "출처" 열을 가릅니다 —
            `naver_movie`는 "네이버 영화", `null`은 "팝콘톡"으로 보입니다.
        movie_id: { type: integer, format: int64 }
        movie: { type: string, description: 영화 제목 }
        rating:
          type: number
          minimum: 0.5
          maximum: 5.0
          multipleOf: 0.5
          example: 4.5
          description: |
            팝콘점수. `dev.reviews.rating`이 `numeric(2,1)`이고
            `CHECK (rating >= 0.5 AND rating <= 5.0)`으로 막습니다.
            admin은 아이콘 5개로 그립니다 — 0.5는 반쪽으로 보입니다.
        content: { type: string }
        status: { $ref: '#/components/schemas/ReviewStatus' }
        created_at: { type: string, format: date }
    ReviewSummary:
      type: object
      required: [total, hidden]
      properties:
        total: { type: integer }
        hidden: { type: integer }

    # ── 회원 ──────────────────────────────────────────────
...